pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bar-assistant-mcp"
version = "0.1.0"
description = "MCP server for Bar Assistant API - manage your home bar shelf"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["mcp", "bar-assistant", "cocktails", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=0.9.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/the-real-py/bar-assistant-mcp"
Repository = "https://github.com/the-real-py/bar-assistant-mcp"
Issues = "https://github.com/the-real-py/bar-assistant-mcp/issues"

[project.scripts]
bar-assistant-mcp = "bar_assistant_mcp.server:main"